We Got Stuff
We’ve got a lot of stuff. The game is really coming along I really
don’t want too much from the game so it’s coming together pretty well.

So far we have a server/client both written in java since node.js
wasn’t necessarily the best at being the authoritative non-browser
server that I wanted to make for this game.

The server is composed of 3(4) main threads. The Server itself
initializes a few variables and then runs simply to continually accept
new clients and assign them an ID and give them a ClientHandler. Since
the server also knows about all of the ClientHandlers, the server can
send messages to all clients.

The second thread the server runs is the ClientHandler. This accepts
input from the client and then parses it to send information back to
the user or to change data locally on the server. The only information
mostly coming from the client is requests to initialize and input
requests.

The WorldHandler thread handles the world to simulate the physics in.
Since doing physics client-side could lead many errors between
instances of the world, I moved all of the world code to just the
server so that the server could have the end-all coordinates of
everybody and then push those coordinates out on a schedule.
WorldHandler also creates all the bodies for the game including the
ships, projectiles and boundaries.

The last thread server-side is the GameEvents thread. This gets loaded
with a EnemyFormation and then begins to deploy enemies after a short
duration. This class (and associated things) needs the most work for
the remainder since right now AI just fly downwards and the current
load out is just three simple enemies.

ClientSide is much simpler. There are two threads. The game rendering
thread which just renders everything at the coordinates that the server
tells it to and the ServerReader thread which simply reads from the
server socket and sends jobs to do in Space.

We’re close to getting projectiles rendered, as right now they just
exist as box2d objects. And then we’ll deal with collisions and then
after collisions, we can work on the AI. After that some menu stuff so
clients can initialize their game better (and maybe create their own
enemy formations) and it’ll be done!